refactor(search): move styles to style
