+++ src/components/main/main.js
@@ -9,7 +9,7 @@ import R from 'ramda'
 import {mux} from 'muxer'
 import Controls from '../controls/controls'
 import Playlist from '../playlist/playlist'
-import SearchBox from '../Search'
+import SearchBox from '../search/search'
 import Empty from '../../lib/RxProxy'
 import * as SC from '../../lib/SoundCloud'
 import css from './main.style'
